+++ ls_UI/modules/unitframes/units/targettarget.lua
@@ -9,11 +9,7 @@ local _G = getfenv(0)
 function UF:CreateTargetTargetFrame(frame)
 	self:CreateSmallUnitFrame(frame)
 
-	local status = frame.TextParent:CreateFontString(nil, "ARTWORK")
-	status:SetFont(GameFontNormal:GetFont(), 16)
-	status:SetJustifyH("LEFT")
-	status:SetPoint("LEFT", frame, "BOTTOMLEFT", 4, -1)
-	frame:Tag(status, "[ls:questicon][ls:sheepicon][ls:phaseicon][ls:leadericon][ls:lfdroleicon][ls:classicon]")
+	frame.Status = self:CreateStatus(frame, frame.TextParent)
 
 	return true
 end